+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Hide sensitive info in screenshots
 - Application settings screen redesign
 - Improve descriptors import UX
+- Watch-only: improved face ID login
 
 ## [5.0.9] - 2025-09-08